+++ src/DialogflowSchema.ts
@@ -151,12 +151,16 @@ export class DialogflowSchema extends Schema {
 
     const language = locale.split("-")[0];
 
-    const agent = _.merge(AGENT, this.mergeManifest(environment), {
-      description: invocationName,
-      language,
-      supportedLanguages,
-      googleAssistant: { project: _.kebabCase(invocationName), startIntents, endIntentIds }
-    });
+    const agent = _.merge(
+      _.cloneDeep(AGENT),
+      _.cloneDeep(this.mergeManifest(environment)),
+      _.cloneDeep({
+        description: invocationName,
+        language,
+        supportedLanguages,
+        googleAssistant: { project: _.kebabCase(invocationName), startIntents, endIntentIds }
+      })
+    );
 
     const file: IFileContent = {
       path: path.join(
@@ -266,6 +270,7 @@ export class DialogflowSchema extends Schema {
     locale = locale.split("-")[0];
     this.builtIntents = intentsByPlatformAndEnvironments.map((rawIntent: IIntent) => {
       let { name, events } = rawIntent;
+      const { webhookForSlotFilling } = rawIntent;
       const { slotsDefinition } = rawIntent;
       name = name.replace("AMAZON.", "");
       const fallbackIntent = name === "FallbackIntent";
@@ -301,7 +306,7 @@ export class DialogflowSchema extends Schema {
         ],
         priority: 500000,
         webhookUsed: true,
-        webhookForSlotFilling: false,
+        webhookForSlotFilling,
         fallbackIntent,
         events
       };