+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed views overwriting existing array
+- Fixed error on publishing information overwriting keys
 
 ## [2.1.1] - 2019-03-29