+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added support for webhookForSlotFilling (dialogflow)
+
 ### Fixed
 
 - Fixed views overwriting existing array
+- Fixed error on publishing information overwriting keys
 
 ## [2.1.1] - 2019-03-29