+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed views overwriting existing array
 - Fixed error on publishing information overwriting keys
+- Filter out empty rows when processing a local spreadsheet
 
 ## [2.1.1] - 2019-03-29