+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed error on publishing information overwriting keys
 - Filter out empty rows when processing a local spreadsheet
 - Fixed empty suggestion chip on VUI
+- Fixed overlap on utterance and slots
 
 ## [2.1.1] - 2019-03-29